MARMOTTA-325 : Deprecate APIs that make it difficult to extend formats
After deprecating both SPARQLGraphResultWriter, and the API methods that
rely on every type of query being applicable to a QueryResultWriter, the
code is substantially simpler, including removal of private methods in
SparqlWebService that are no longer required.

QueryResultWriter is only relevant to Tuple and Boolean results, so
hacking it to work with Graph results instead of using RDFHandler makes
it difficult to integrate with other API methods. Referencing RDFHandler
isn't even necessary in SparqlService if you only want to output to a
document, as
QueryResultIO.write(GraphQueryResult,RDFFormat,OutputStream) does the
resolution and handling automatically, as long as you can deal with
RDFHandlerException.