MARMOTTA-325 : Make io service more flexible wrt dynamic services
If parsers were loaded into the registry after startup, the current code
would not recognise them. In order to make it possible to do this for
marmotta, the io service needs to call the authoritative service
registry when asked about questions.

In addition, the list of accept types could contain duplicates if more
than one parser supports a given format, so filtering through a set.