MARMOTTA-325 : Add hack to remove application/xml
There is no theoretical reason why we can't respond with application/xml
if they send application/xml, but in order to distinguish tuple and
graph results it may be easier just to remove it. This makes the
specific test that calls for application/sparql-results+xml as the
result for application/xml to succeed. Without the hack they get what
they asked for, application/xml