Add statements and compare size before commit.
There was an error while adding two equally statements, but do not commit last addition. The result should be 1, but instead, show 2. This was because method getSize() returned database size + a batch, and because these two statements were the same, result was 2 instead of 1.

Solution: make a flushBatch before each call to getSize() method.

Signed-off-by: Gustavo.