Modify query to verify statements with null values.
Modification of select query in order to check if the statement exists. This was failing due that value=null is not valid in SQL it should be checked as value is null when a null value is passed.

More info about the agreement to modify query: https://lists.apache.org/list.html?dev@marmotta.apache.org

Signed-of-by: Gustavo.